Migrate elzilla to async and make comment posting a mode
Moving to elfuture required adding one extra elfuture operation called
elfuture-join, which waits for a group of futures and invokes a new future with
a list of their results. This makes it convenient to invoke multiple URLs at
once and get all their responses as one batch.

The comment posting mode is fairly basic, with C-c C-c and C-c C-k being the
only bindings to post the comment/cancel the posting.